requirements.txt
# Core ML & Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Explainability
shap==0.44.0

# Web Framework
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# File handling
openpyxl==3.1.2
python-multipart==0.0.6

# PDF Generation
reportlab==4.0.7
fpdf2==2.7.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
requests==2.31.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0

# Type hints
typing-extensions==4.9.0

# Authentication (optional for production)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
